Fix various integer-formattingr-elated issues
- format string in MinGW (or rather the used version of the MS C
  runtime) may not support about %z for size_t: cast to uint64_t
- ERROR_SUCCESS may be predefined: cast it to int

+++ tests/test-api.c
@@ -469,7 +469,7 @@ void test_scanner()
     yr_scanner_destroy(scanner1);
     yr_scanner_destroy(scanner2);
     yr_rules_destroy(rules);
-    printf("expecting ERROR_SUCCESS (%d), got: %d\n", ERROR_SUCCESS, result);
+    printf("expecting ERROR_SUCCESS (%d), got: %d\n", (int)ERROR_SUCCESS, result);
     exit(EXIT_FAILURE);
   }
